requirements.txt
flask
tensorflow==2.15.0
pandas
numpy
nltk 
langdetect
gTTS
streamlit
bnlp-toolkit
